Documents/uni/321/tutoreal 2/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
MarkupSafe
nltk
transformers
eventlet